fix rename AddToFavoriteSong
