+++ src/Application/Command/AddToFavoriteSong/AddToFavoriteSong.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Application\Command\AddToFavoriteSong;
+
+class AddToFavoriteSong
+{
+    public int $userId;
+    public int $songId;
+}
