--- src/Application/Command/AddToFavoriteUser/AddToFavoriteUser.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App\Application\Command\AddToFavoriteUser;
-
-class AddToFavoriteUser
-{
-    public int $userId;
-    public int $songId;
-}
